refactor: Move custom logic to hook
